LayoutTests/accessibility/ios-simulator/ancestor-computation-expected.txt

This test ensures accessibility elements properly track their ancestor types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Verifying {"isInDescriptionListDetail":"true"} for all descendants of outer-group.
PASS element.isInDescriptionListDetail is true
PASS element.isInDescriptionListDetail is true
PASS element.isInDescriptionListDetail is true
Finished verifying {"isInDescriptionListDetail":"true"} for all descendants of outer-group.

Verifying {"isInDescriptionListDetail":"false","isInDescriptionListTerm":"true"} for all descendants of outer-group.
PASS element.isInDescriptionListDetail is false
PASS element.isInDescriptionListTerm is true
PASS element.isInDescriptionListDetail is false
PASS element.isInDescriptionListTerm is true
PASS element.isInDescriptionListDetail is false
PASS element.isInDescriptionListTerm is true
Finished verifying {"isInDescriptionListDetail":"false","isInDescriptionListTerm":"true"} for all descendants of outer-group.

Verifying {"isInDescriptionListTerm":"false","isInCell":"false"} for all descendants of outer-group.
PASS element.isInDescriptionListTerm is false
PASS element.isInCell is false
PASS element.isInDescriptionListTerm is false
PASS element.isInCell is false
PASS element.isInDescriptionListTerm is false
PASS element.isInCell is false
Finished verifying {"isInDescriptionListTerm":"false","isInCell":"false"} for all descendants of outer-group.

Verifying {"isInCell":"true"} for all descendants of outer-group.
PASS element.isInCell is true
PASS element.isInCell is true
PASS element.isInCell is true
Finished verifying {"isInCell":"true"} for all descendants of outer-group.

Verifying {"isInCell":"false"} for all descendants of outer-group.
PASS element.isInCell is false
PASS element.isInCell is false
PASS element.isInCell is false
Finished verifying {"isInCell":"false"} for all descendants of outer-group.

Verifying {"isInCell":"true"} for all descendants of outer-group.
PASS element.isInCell is true
PASS element.isInCell is true
PASS element.isInCell is true
Finished verifying {"isInCell":"true"} for all descendants of outer-group.

Verifying {"isInDescriptionListDetail":"false","isInDescriptionListTerm":"false","isInCell":"false"} for all descendants of outer-group.
PASS element.isInDescriptionListDetail is false
PASS element.isInDescriptionListTerm is false
PASS element.isInCell is false
PASS element.isInDescriptionListDetail is false
PASS element.isInDescriptionListTerm is false
PASS element.isInCell is false
PASS element.isInDescriptionListDetail is false
PASS element.isInDescriptionListTerm is false
PASS element.isInCell is false
Finished verifying {"isInDescriptionListDetail":"false","isInDescriptionListTerm":"false","isInCell":"false"} for all descendants of outer-group.

PASS successfullyParsed is true

TEST COMPLETE
